Add ability to edit task name
Users can edit the current task name. change_task_name() has been defined and added into check_input(). changed the order of the function definitions a bit to match the current order. Need to add an exec file to this. 

The user should have the ability to say something like "I've been done with that task, and have actually been working on this other task. Here's the name and how long I've been working on that: __". 

Basically, the ability to end a task, subtract from the current task time, add to the current task time, create another task, and add to that time.