Manual task enter feature
The user can type in "manual" and is prompted to enter a task name and time. Time is converted to seconds and the user is brought back to the beginning of the loop. Things are messy right now. There are a few too many uses of `global` and the flow is not very readable. At this point, barring some smaller changes, what the code needs is some review and refactoring. Classes might be needed.